Struct Type & Editor/Window by Jahorta (#181)
* Implement Struct Type and create Editor/Window for Struct editing/creation

+++ Source/GUI/MemWatcher/MemWatchWidget.cpp
@@ -247,10 +247,15 @@ void MemWatchWidget::onMemWatchContextMenuRequested(const QPoint& pos)
       connect(unlockSelection, &QAction::triggered, this, [this] { onLockSelection(false); });
       contextMenu->addAction(unlockSelection);
       contextMenu->addSeparator();
-      QAction* const editValue{new QAction(tr("Edit Value"), this)};
-      connect(editValue, &QAction::triggered, this, [this, index]() { m_watchView->edit(index); });
-      contextMenu->addAction(editValue);
-      contextMenu->addSeparator();
+
+      if (!GUICommon::isContainerType(entry->getType()))
+      {
+        QAction* const editValue{new QAction(tr("Edit Value"), this)};
+        connect(editValue, &QAction::triggered, this,
+                [this, index]() { m_watchView->edit(index); });
+        contextMenu->addAction(editValue);
+        contextMenu->addSeparator();
+      }
     }
   }
   else
@@ -261,13 +266,26 @@ void MemWatchWidget::onMemWatchContextMenuRequested(const QPoint& pos)
   const QModelIndexList simplifiedSelection{simplifySelection()};
   if (!simplifiedSelection.empty())
   {
-    QAction* const groupAction{new QAction(tr("&Group"), this)};
-    connect(groupAction, &QAction::triggered, this, &MemWatchWidget::groupCurrentSelection);
-    contextMenu->addAction(groupAction);
-    contextMenu->addSeparator();
+    bool canGroupSelection = true;
+    for (const QModelIndex& selectedIndex : simplifiedSelection)
+    {
+      const MemWatchTreeNode* selectedNode = m_watchModel->getTreeNodeFromIndex(selectedIndex);
+      if (selectedNode != nullptr && selectedNode->getParent() != nullptr &&
+          (!selectedNode->getParent()->isGroup() &&
+           selectedNode->getParent() != m_watchModel->getRootNode() &&
+           GUICommon::isContainerType(selectedNode->getParent()->getEntry()->getType())))
+        canGroupSelection = false;
+    }
+    if (canGroupSelection)
+    {
+      QAction* const groupAction{new QAction(tr("&Group"), this)};
+      connect(groupAction, &QAction::triggered, this, &MemWatchWidget::groupCurrentSelection);
+      contextMenu->addAction(groupAction);
+      contextMenu->addSeparator();
+    }
   }
 
-  if (!node || node->isGroup())
+  if (node == m_watchModel->getRootNode() || node->isGroup())
   {
     QAction* const addGroup{new QAction(tr("Add gro&up"), this)};
     connect(addGroup, &QAction::triggered, this, &MemWatchWidget::onAddGroup);
@@ -323,10 +341,15 @@ void MemWatchWidget::onMemWatchContextMenuRequested(const QPoint& pos)
   connect(paste, &QAction::triggered, this, [this, index] { pasteWatchFromClipBoard(index); });
   contextMenu->addAction(paste);
 
-  contextMenu->addSeparator();
   QAction* deleteSelection = new QAction(tr("&Delete"), this);
-  connect(deleteSelection, &QAction::triggered, this, [this] { onDeleteSelection(); });
-  contextMenu->addAction(deleteSelection);
+  if (node != m_watchModel->getRootNode() &&
+      !(node->getParent() != nullptr && node->getParent()->getEntry() &&
+        GUICommon::isContainerType(node->getParent()->getEntry()->getType())))
+  {
+    contextMenu->addSeparator();
+    connect(deleteSelection, &QAction::triggered, this, [this] { onDeleteSelection(); });
+    contextMenu->addAction(deleteSelection);
+  }
 
   QModelIndexList selection = m_watchView->selectionModel()->selectedRows();
   if (selection.count() == 0)
@@ -427,7 +450,7 @@ void MemWatchWidget::pasteWatchFromClipBoard(const QModelIndex& referenceIndex)
   {
     const QString nodeStr{QApplication::clipboard()->text()};
     const QJsonDocument loadDoc{QJsonDocument::fromJson(nodeStr.toUtf8())};
-    copiedRootNode.readFromJson(loadDoc.object(), nullptr);
+    copiedRootNode.readFromJson(loadDoc.object(), QMap<QString, QString>());
   }
 
   const QVector<MemWatchTreeNode*> children{copiedRootNode.getChildren()};
@@ -444,22 +467,32 @@ void MemWatchWidget::onWatchDoubleClicked(const QModelIndex& index)
   if (index != QVariant())
   {
     MemWatchTreeNode* node = static_cast<MemWatchTreeNode*>(index.internalPointer());
-    if (index.column() == MemWatchModel::WATCH_COL_TYPE && !node->isGroup())
+
+    if (node->isGroup() || (node->getParent() && node->getParent()->getEntry() &&
+                            GUICommon::isContainerType(node->getParent()->getEntry()->getType())))
+      return;
+
+    if (index.column() == MemWatchModel::WATCH_COL_TYPE)
     {
       MemWatchEntry* entry = node->getEntry();
       int typeIndex = static_cast<int>(entry->getType());
-      DlgChangeType* dlg = new DlgChangeType(this, typeIndex, entry->getLength());
+      DlgChangeType* dlg =
+          new DlgChangeType(this, typeIndex, entry->getLength(), m_structDefs->getStructNames(),
+                            entry->getStructName());
       if (dlg->exec() == QDialog::Accepted)
       {
         Common::MemType theType = static_cast<Common::MemType>(dlg->getTypeIndex());
+        if (theType == Common::MemType::type_struct && dlg->getStructName() != QString())
+          entry->setStructName(dlg->getStructName());
+
         m_watchModel->changeType(index, theType, dlg->getLength());
         m_hasUnsavedChanges = true;
       }
     }
-    else if (index.column() == MemWatchModel::WATCH_COL_ADDRESS && !node->isGroup())
+    else if (index.column() == MemWatchModel::WATCH_COL_ADDRESS)
     {
       MemWatchEntry* entryCopy = new MemWatchEntry(node->getEntry());
-      DlgAddWatchEntry dlg(false, entryCopy, this);
+      DlgAddWatchEntry dlg(false, entryCopy, m_structDefs->getStructNames(), this);
       if (dlg.exec() == QDialog::Accepted)
       {
         m_watchModel->editEntry(dlg.stealEntry(), index);
@@ -583,7 +616,7 @@ void MemWatchWidget::onAddGroup()
 
 void MemWatchWidget::onAddWatchEntry()
 {
-  DlgAddWatchEntry dlg(true, nullptr, this);
+  DlgAddWatchEntry dlg(true, nullptr, m_structDefs->getStructNames(), this);
   if (dlg.exec() == QDialog::Accepted)
   {
     addWatchEntry(dlg.stealEntry());
@@ -696,6 +729,17 @@ void MemWatchWidget::onRowsInserted(const QModelIndex& parent, const int first, 
                                  lastIndex.siblingAtColumn(MemWatchModel::WATCH_COL_NUM - 1)};
 
   QItemSelectionModel* const selectionModel{m_watchView->selectionModel()};
+  // If the parent node is a container and it is not expanded, do not select the child node or
+  // expand it.
+  const MemWatchTreeNode* parentNode = MemWatchModel::getTreeNodeFromIndex(parent);
+  if (parentNode != nullptr && !parentNode->isGroup() &&
+      parentNode != m_watchModel->getRootNode() &&
+      GUICommon::isContainerType(parentNode->getEntry()->getType()))
+  {
+    selectionModel->clearSelection();
+    return;
+  }
+
   selectionModel->select(selection, QItemSelectionModel::ClearAndSelect);
   selectionModel->setCurrentIndex(lastIndex, QItemSelectionModel::Current);
 
@@ -718,6 +762,9 @@ void MemWatchWidget::onCollapsed(const QModelIndex& index)
     return;
   node->setExpanded(false);
   m_hasUnsavedChanges = true;
+
+  if (!node->isGroup() && GUICommon::isContainerType(node->getEntry()->getType()))
+    m_watchModel->collapseContainerNode(node);
 }
 
 void MemWatchWidget::onExpanded(const QModelIndex& index)
@@ -727,6 +774,9 @@ void MemWatchWidget::onExpanded(const QModelIndex& index)
     return;
   node->setExpanded(true);
   m_hasUnsavedChanges = true;
+
+  if (!node->isGroup() && GUICommon::isContainerType(node->getEntry()->getType()))
+    m_watchModel->expandContainerNode(node);
 }
 
 QTimer* MemWatchWidget::getUpdateTimer() const
@@ -753,17 +803,25 @@ void MemWatchWidget::openWatchFile(const QString& fileName)
   }
   if (!srcFileName.isEmpty())
   {
+    bool structsOnly = false;
+    bool clearStructTree = false;
     if (m_watchModel->hasAnyNodes())
     {
       QMessageBox* questionBox = new QMessageBox(
           QMessageBox::Question, "Asking to merge lists",
           "The current watch list has entries in it, do you want to merge it with this one?",
           QMessageBox::Cancel | QMessageBox::No | QMessageBox::Yes, this);
+      QPushButton* structOnly = questionBox->addButton(tr("Structs only"), QMessageBox::AcceptRole);
       int answer = questionBox->exec();
       if (answer == QMessageBox::Cancel)
         return;
-      if (answer == QMessageBox::No)
+      else if (answer == QMessageBox::No)
+      {
         m_watchModel->clearRoot();
+        clearStructTree = true;
+      }
+      else if (answer != QMessageBox::Yes && questionBox->clickedButton() == structOnly)
+        structsOnly = true;
     }
 
     QFile watchFile(srcFileName);
@@ -786,7 +844,10 @@ void MemWatchWidget::openWatchFile(const QString& fileName)
     QByteArray bytes = watchFile.readAll();
     watchFile.close();
     QJsonDocument loadDoc(QJsonDocument::fromJson(bytes));
-    m_watchModel->loadRootFromJsonRecursive(loadDoc.object());
+    QMap<QString, QString> structReplacements = {};
+    emit loadStructDefsFromJson(loadDoc.object(), structReplacements, clearStructTree);
+    if (!structsOnly)
+      m_watchModel->loadRootFromJsonRecursive(loadDoc.object(), structReplacements);
     updateExpansionState();
     m_watchListFile = srcFileName;
     m_hasUnsavedChanges = false;
@@ -809,6 +870,7 @@ bool MemWatchWidget::saveWatchFile()
     }
     QJsonObject root;
     m_watchModel->writeRootToJsonRecursive(root);
+    emit writeStructDefTreeToJson(root);
     QJsonDocument saveDoc(root);
     watchFile.write(saveDoc.toJson());
     watchFile.close();
@@ -838,6 +900,7 @@ bool MemWatchWidget::saveAsWatchFile()
     }
     QJsonObject root;
     m_watchModel->writeRootToJsonRecursive(root);
+    emit writeStructDefTreeToJson(root);
     QJsonDocument saveDoc(root);
     watchFile.write(saveDoc.toJson());
     watchFile.close();
@@ -991,6 +1054,30 @@ QString MemWatchWidget::saveWatchModel()
   m_watchModel->writeRootToJsonRecursive(root);
   QJsonDocument saveDoc(root);
   return saveDoc.toJson();
+}
+
+void MemWatchWidget::setStructDefs(StructTreeNode* structDefs, QMap<QString, StructDef*> structMap)
+{
+  m_structDefs = structDefs;
+  m_watchModel->setStructMap(structMap);
+}
+
+void MemWatchWidget::onUpdateStructDetails(QString structName)
+{
+  m_watchModel->updateStructEntries(structName);
+  m_hasUnsavedChanges = true;
+}
+
+void MemWatchWidget::onUpdateStructName(QString oldName, QString newName)
+{
+  m_watchModel->onStructNameChanged(oldName, newName);
+  m_hasUnsavedChanges = true;
+}
+
+void MemWatchWidget::onStructDefAddRemove(QString structName, StructDef* structDef)
+{
+  m_watchModel->onStructDefAddRemove(structName, structDef);
+  m_hasUnsavedChanges = true;
 }
 
 void MemWatchWidget::updateExpansionState(const MemWatchTreeNode* const node)
